requirements.txt
sqlalchemy
google-generativeai